nxp-s32g: atf: update compatible with SDK BSP44
Issue: LINUXEXEC-34820

The SDK BSP44 has been released, this patch updates atf recipe to SDK
BSP44 release.

In detail, there is a new introduced variable "DDR_FW_PATH", which is
the location where the DDR firmwares are put, it is a required compile
option in SDK BSP44.

And the "0001-s32-clk-Return-the-preset-freq-when-we-can-t-calcula.patch"
patch is updated by adding "Upstream-Status" to fix patch QA issue.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>