nxp-s32g: optee-os: add v4.0.0 support for S32G platform
Issue: LIN1025-2250

The latest optee version in LTS25 has been upgraded to v4.4.0, while it is
not compatible with S32G platform, which still needs the v4.0.0 version to
make optee feature workable, especially for HSE support.

So, add v4.0.0 optee back for S32G platform.
In this patch, the optee-os related recipe files are added, and these files
refer to SDK BSP44.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>