nxp-s32g: atf: use /dev/null as default ddr firmware path to make the defaul build pass
Issue: LIN1025-2375

Since BSP44 introduced a standalone DDR firmware, the auto build will fail
if the path of DDR firmware not set. So let's use /dev/null as default one
to make auto build pass, and specify the DDR_FW_PATH to indicate the real
used firmware to make the bsp workable. Moreover, there will be a error log
if the DDR_FW_PATH set to a invalid path.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>