nxp-s32g: README: add a PCIe known issue description
Issue: LIN1025-2254

For S32G platform, there will be following issue if NVMe SSD card plug-in:
    s32cc-pcie 40400000.pcie: PHY error detected after Hot-Plug event

This is caused due to lack of AER support in PCIe driver, which is a known
issue, and vendor says it will be fixed in the future release. So add
description for this issue.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>