nxp-s32g: atf: update mbedtls makefile to update TF_CFLAGS to fix redundant-decls error
Issue: LIN1025-3112

When atf is compiled in secure mode, it depends on mbedtls. Recently, mbedtls
has been updated, and there is a commit 3ea3635d2f46 ("Move the inclusion of
crypto_sizes.h and crypto_struct.h in crypto.h"), which modifies the
include/psa/crypto.h of mbedtls, and this will cause the following atf
compile error:
  include/psa/crypto.h:122:29: error: redundant redeclaration of 'psa_key_attributes_init' [-Werror=redundant-decls]

But for psa APIs, there indeed functions are declared in multiple header files,
so remove the redundant-decls warning as error setting to fix the above issue.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>